% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aes_puller.R
\name{aes_puller}
\alias{aes_puller}
\title{retrieve a vector storing values of a specified aestetich}
\usage{
aes_puller(plot_object, n_of_layers, required_aes)
}
\arguments{
\item{plot_object}{the ggplot object being analysed}

\item{n_of_layers}{number of layers within the ggplot object}

\item{required_aes}{the ggplot object aestetich required}
}
\value{
a vector containing values of the variable the required specified aestetich is mapped to
}
\description{
retrieve a vector storing values of a specified aestetich
}
