% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordinates_lister.R
\name{coordinates_lister}
\alias{coordinates_lister}
\title{list coordinates layers in a ggplot object}
\usage{
coordinates_lister(plot_object)
}
\arguments{
\item{plot_object}{the ggplot object being analysed}
}
\value{
a vector storing all coordinates found within layers of a ggplot object
}
\description{
list coordinates layers in a ggplot object
}
