% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cozy_plot.R
\name{cozy_plot}
\alias{cozy_plot}
\title{checking for the presence of overplotting within a ggplot object.}
\usage{
cozy_plot(plot_object, n_of_layers, overplotting_floor, aes_db)
}
\arguments{
\item{plot_object}{the ggplot object being analysed}

\item{n_of_layers}{number of layers within the ggplot object}

\item{overplotting_floor}{a custom threshold of distance among points to detect overplotting}

\item{aes_db}{a data frame storing plot object aestetichs and variable mapped against them}
}
\value{
a list containing an overall test result with TRUE in presence of overplotting and the median distance observed between points
}
\description{
checking for the presence of overplotting within a ggplot object.
}
