% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/does_it_need_geom_smooth.R
\name{does_it_need_geom_smooth}
\alias{does_it_need_geom_smooth}
\title{study the relationship among variables to find out if a geom_smooth is needed within the plot}
\usage{
does_it_need_geom_smooth(plot_object, n_of_layers, correlation_threshold,
  aes_db)
}
\arguments{
\item{plot_object}{the ggplot object being analysed}

\item{n_of_layers}{number of layers within the ggplot object}

\item{correlation_threshold}{a judgmentally specified threshold to test for the need of a trend line on the plot}

\item{aes_db}{a data frame storing plot object aestetichs and variable mapped against them}
}
\value{
a list storing in two separate vectors the final result of the test and the observed distance correlation.
the test outcome is TRUE in case a trend line is deemed necessary.
}
\description{
\code{does_it_need_geom_smooth} computes distance correlation among x and y aestetich to evaluate it against a judgmental threshold.
if this threshold is breached and no \code{geom_smooth} is observed among ggplot objects layers a postive outcome is produced for the test.
}
