% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filled_barplot.R
\name{filled_barplot}
\alias{filled_barplot}
\title{test a ggplot object to understand if it represents a barplot with filled bars}
\usage{
filled_barplot(plot_object, n_of_layers)
}
\arguments{
\item{plot_object}{the ggplot object being analysed}

\item{n_of_layers}{number of layers within the ggplot object}
}
\value{
returns TRUE when the barplot is composed of filled bars and FALSE in the opposite case.
}
\description{
this function tests a ggplot object to understand if it represents a barplot with filled bars.
Based on data visualization literature this kind of plots tend to reduce the **data to ink ratio** meaning the amount of data showed related to amount of ink employed.
Lower values of this ratio are related to non-effective plots.
A control is performed to exclude we are simply looking at a pie chart based on an intermediate \code{geom_bar} layer
}
\references{
The Visual Display of Quantitative Information, E.R.Tufte, GRaphics Press . Cheshire, Connecticut
}
