% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms_lister.R
\name{geoms_lister}
\alias{geoms_lister}
\title{list geoms in a ggplot object}
\usage{
geoms_lister(plot_object, n_of_layers)
}
\arguments{
\item{plot_object}{the ggplot object being analysed}

\item{n_of_layers}{number of layers within the ggplot object}
}
\value{
a vector listing all geoms found within the given ggplot object
}
\description{
list geoms in a ggplot object
}
