% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heavy_background.R
\name{heavy_background}
\alias{heavy_background}
\title{look for heavy background on a ggplot plot}
\usage{
heavy_background(plot_object)
}
\arguments{
\item{plot_object}{the ggplot object being analysed}
}
\value{
returns TRUE if the plot is depicted on an heavy background, FALSE if the background is white or transparent
}
\description{
this function tests a ggplot object to understand if it represents a plot with a background neither transparent nor white.
Based on data visualization literature this kind of plots tend to reduce the **data to ink ratio** meaning the amount of data showed related to amount of ink employed.
Lower values of this ratio are related to non-effective plots.
}
\references{
The Visual Display of Quantitative Information, E.R.Tufte, GRaphics Press . Cheshire, Connecticut
}
