% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histogram_bins_tester.R
\name{histogram_bins_tester}
\alias{histogram_bins_tester}
\title{compare the actual bins width against the Freedman Diaconis rule}
\usage{
histogram_bins_tester(plot_object, n_of_layers, default_n_of_bins)
}
\arguments{
\item{plot_object}{the ggplot object being analysed}

\item{n_of_layers}{number of layers within the ggplot object}

\item{default_n_of_bins}{the default number of bins set by ggplot2. employed if no custom bin was set from the user within the ggplot object}
}
\value{
an object storing the general test result, the optimal binwidth based on Freedman Diaconis and the distance of the actual binwidth from this measure.
TRUE is produced as a result if the actual width is not optimal.
}
\description{
this function computes the optimal bins width based on the Freedman Diaconis rule and compares it against the actual bin width set within the ggplot object.
}
