% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_horizontal_barplot.R
\name{is_horizontal_barplot}
\alias{is_horizontal_barplot}
\title{a simple function to test for the presence of too many dimension for a 2D plotting space}
\usage{
is_horizontal_barplot(plot_object, n_of_layers)
}
\arguments{
\item{plot_object}{the ggplot object being analysed}

\item{n_of_layers}{number of layers within the ggplot object}
}
\value{
TRUE in case the plot object is a bar plot and it is horizontally layed, FALSE in the opposite case.
}
\description{
a simple function to test for the presence of too many dimension for a 2D plotting space
}
