% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_pie_chart.R
\name{is_pie_chart}
\alias{is_pie_chart}
\title{test a ggplot object to understand if it is a pie chart}
\usage{
is_pie_chart(plot_object, n_of_layers)
}
\arguments{
\item{plot_object}{the ggplot object being analysed}

\item{n_of_layers}{number of layers within the ggplot object}
}
\value{
TRUE if the plot is a pie chart and FALSE in the opposite case
}
\description{
to understand if aplot is a pie chart this functions look for the presence of a \code{geom_bar} layer and a \code{coord_polar} layer.
}
