% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels_finder.R
\name{labels_finder}
\alias{labels_finder}
\title{look for a specific label within a ggplot object}
\usage{
labels_finder(plot_object, searched_label)
}
\arguments{
\item{plot_object}{the ggplot object being analysed}

\item{searched_label, }{among title, subtitle and caption}
}
\value{
TRUE if the searched label was found, FALSE in the opposite case
}
\description{
look for a specific label within a ggplot object
}
