% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/labels_reader.R
\name{labels_reader}
\alias{labels_reader}
\title{look for special charaters in a plot labels}
\usage{
labels_reader(plot_object)
}
\arguments{
\item{plot_object}{the ggplot object being analysed}
}
\value{
a list storing test result and the list of special characters found within labels.
}
\description{
this function match the labels of the given ggplot object against a set of predefined special characters
}
