% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappings_lister.R
\name{mappings_lister}
\alias{mappings_lister}
\title{storing aestetich variables in a dataframe}
\usage{
mappings_lister(plot_object, n_of_layers)
}
\arguments{
\item{plot_object}{the ggplot object being analysed}

\item{n_of_layers}{number of layers within the ggplot object}
}
\value{
a data frane storing for each aestetich the mapped variable
}
\description{
storing aestetich variables in a dataframe
}
