% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_reader.R
\name{metadata_reader}
\alias{metadata_reader}
\title{draw a data frame containing relevant metadata of a ggplot object}
\usage{
metadata_reader(plot_object = NULL)
}
\arguments{
\item{plot_object}{the ggplot object being analysed}
}
\value{
a list storing results and additional data related to each test performed.
}
\description{
\code{metadata_reader} function performs a number of test on the provided ggplot object to draw some intelligence about it.
It tries to derive attributes ofthe plot relate to four data visualization theory areas:
- the adequateness of labelling
- the data density
- the data to ink ratio
- the readability of the plot
The results of performed tests are stored in a structued list object, so to be provided as an input to a gradient boosting model.
See \code{scorer} function documentation for more info about this subsequent step.
}
