% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_labels.R
\name{outlier_labels}
\alias{outlier_labels}
\title{check for the presence of outliers not labelled within a ggplot object}
\usage{
outlier_labels(plot_object, n_of_layers, p_build, aes_db)
}
\arguments{
\item{plot_object}{the ggplot object being analysed}

\item{n_of_layers}{number of layers within the ggplot object}

\item{p_build}{an object obtained running \code{ggplot_build} on the ggplot object}

\item{aes_db}{a data frame storing plot object aestetichs and variable mapped against them}
}
\value{
a list storing a TRUE in case there are no outliers or the detected outliers are labelled and FALSE in the other cases.
}
\description{
for x and y variable the Tukey outlier rule is applied to evaluate the presence of outliers.
If any outlier is identified one more check is performed to retrieve among plot metadata the presence of some annotation related to that data.
}
