% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorer.R
\name{scorer}
\alias{scorer}
\title{scoring the effectivenes of ggplot object applying a gradient boosting model}
\usage{
scorer(plot_metadata = NULL)
}
\arguments{
\item{plot_metadata}{a data frame storing ggplot object metadata, resulting from a previous call of \code{metadata_reader} function.}
}
\value{
a numeric value
}
\description{
\code{scorer} function handles the running of a gradient boosting pre-trained model on the metadata of a ggplot object.
As a result a 0 to 1 score is produced, measuring the effectiveness of the related plot based on a set of rules drawn from data visualization theory.
}
