% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scorer_bot.R
\name{scorer_bot}
\alias{scorer_bot}
\title{produce a customised report analysing the submitted ggplot object}
\usage{
scorer_bot(plot_object = NULL)
}
\arguments{
\item{plot_object}{the ggplot object being analysed}
}
\value{
a reveljs deck of slide containing an evaluation of the effectiveness of ggplot objects and suggestion about how to improve it.
}
\description{
call to \code{scorer_bot} function produces the sequential calling of all function included within
the \code{vizscorer} package, with the purpose to produce a customised report about the submitted ggplot object. The function handles call to \code{metadata_reader} function
to draw metadata from the plot object in a format ready to be submitted to the scorer algorithm. within the next step those metadata are submitted to the \code{scorer} function which apply a gradient boosting pre-trained model to them.
a final score from 0 to 1 is computed as a probability to be a good plot.
To provide more relevant information to the user a set of advices are retrieved from a built-in knowledge base, selecting only those relevant to suboptimal areas of the plot, observed from metadata.
Finally a revealjs deck of slides is programmatically created leveraging a quasi-real natural language generation algorithm.
}
