% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_for_geom_bar.R
\name{test_for_geom_bar}
\alias{test_for_geom_bar}
\title{look for geom_bar layer on a plot}
\usage{
test_for_geom_bar(plot_object, n_of_layers)
}
\arguments{
\item{plot_object}{the ggplot object being analysed}

\item{n_of_layers}{number of layers within the ggplot object}
}
\value{
TRUE if a geom_bar is found within ggplot object layers, FALSE in the opposite case.
}
\description{
look for geom_bar layer on a plot
}
