% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_for_histogram.R
\name{test_for_histogram}
\alias{test_for_histogram}
\title{analyse a ggplot object to understand if it is an histogram}
\usage{
test_for_histogram(plot_object, n_of_layers)
}
\arguments{
\item{plot_object}{the ggplot object being analysed}

\item{n_of_layers}{number of layers within the ggplot object}
}
\value{
TRUE in case the plot is an histogram, FALSE in the opposite case
}
\description{
this function checks two conditions to understand if the provided ggplot object represent an histogram:
- presence of a geom_bar layer
- presence of bindwidth parameter among parameters of that layer
}
