% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/too_many_dimensions.R
\name{too_many_dimensions}
\alias{too_many_dimensions}
\title{a simple function to test for the presence of too many dimension for a 2D plotting space}
\usage{
too_many_dimensions(plot_object, n_of_layers)
}
\arguments{
\item{plot_object}{the ggplot object being analysed}

\item{n_of_layers}{number of layers within the ggplot object}
}
\value{
TRUE if the plot contains too many dimension, FALSE in the opposite case.
}
\description{
this function checks for the presence of both point or line and the mapping of some variable on the \code{size} aestetich.
This check is based on data visualization principle requesting that the number of dimension depicted should not be higher thant the number of dimension of the data.
}
\references{
The Visual Display of Quantitative Information, E.R.Tufte, GRaphics Press . Cheshire, Connecticut
}
